Create i8255Channel

+++ lib/utils/utils.cpp
@@ -265,7 +265,7 @@ std::string util_entry(std::string crunched, size_t fileSize, bool is_dir, bool 
             "%-8s %-3s %10lu  %2u-%02u-%02u  %2u:%02u%c",
             basename.c_str(),
             ext.c_str(),
-            fileSize,
+            (long unsigned int) fileSize,
             month,
             day,
             year,
